## [0.8.0](https://pypi.org/project/bento-cli/0.8.0/) - 2020-01-24

This release represents a major shift in Bento's default behavior: It emphasizes an incremental
and personal, rather than team-wide, workflow that makes Bento a smaller commitment to use:

1. Other project contributors won’t see Bento files or have their workflows changed.
2. You no longer need to manually run Bento. After initialization Bento will automatically
   check for issues in your code as you commit, analyzing only the files that have changed.
3. You won’t see a project’s old issues (tech debt) during initialization. To view them,
   run

   ```bash
   bento check --all
   ```

### Migration

Project configurations have changed in version 0.8. In order to migrate a project from
version 0.7 or earlier:

- Ensure Bento has been upgraded using `pip3 install --upgrade bento-cli`. Run `bento --version`
  to validate Bento’s version.
- `rm -r .bento*` (Use `git rm` if you have previously added Bento files to source control).
- Run `bento init` in the project root.

### New requirements

- Docker must be installed, and the docker client running, to use Bento.

### Changed

How you use Bento has changed significantly in 0.8.

#### Usage changes

- `bento check` will now only check staged changes.
  - Use `bento check --all` to check the entire project.
- `bento archive` will archive findings due to staged diff:

  - Use `bento archive --all` to archive all findings in a project.

#### Other changes

- `hadolint` and `shellcheck` are now enabled by default.
- Messages for findings are no longer truncated.
- `bento init` will now install an empty configuration on a project it can not identify;
  tools may then be manually enabled using `bento enable tool TOOL`. Use `bento enable tool --help`
  to list tools.
- Virtual environments for Python tools are now installed in your home directory, instead of in your
  project directories.
- `eslint` is now installed in your project’s `.bento` directory, and will not modify your project’s
  `package.json`.
- Tool and check names have been modified to improve readability.
- `eslint` is disabled by default. To enable it run `bento enable tool eslint`.

### Added

- On init, Bento will prompt you to ask if you want to add ignore patterns to your global Git ignore file.
  If you agree, Bento will alter this file to ignore Bento configuration files in your git projects.
- `bento enable autorun` and `bento disable autorun` will cause Bento to either begin or stop analyzing
  code on every commit.
- Tab completion can now be installed by Bento. To install for your shell, run `bento enable autocomplete`.
  To remove tab completion, run `bento disable autocomplete`.

### Removed

- `bento check --show-all` has been removed. Use `bento check --all` instead. Archived findings can be found at `~/.bento/archive.json`.
- `bento install-hook` has been removed. Use `bento enable autorun` to run Bento on every commit.
- The histogram formatter is no longer used by default. To show findings with a histogram, run
  `bento check -f histo`.

+++ tests/acceptance/qa.py
@@ -1,75 +1,96 @@
 import os
+import re
 import subprocess
+import sys
 import tempfile
 from pathlib import Path
-from typing import Any, Callable, List, Optional, Union
+from typing import Any, Callable, List, Mapping, Optional
 
 import yaml
 
 import bento.constants
+from _pytest.python import Metafunc
 
-Expectation = Union[str, List[str]]
+ALL_TESTS = [
+    d.name
+    for d in Path(__file__).parent.resolve().iterdir()
+    if d.is_dir() and (d / "commands.yaml").exists()
+]
 
-BENTO_REPO_ROOT = str(Path(__file__).parent.parent.parent)
+BENTO_REPO_ROOT = str(Path(__file__).parent.parent.parent.resolve())
+DYNAMIC_SECONDS = re.compile(r"([\s\S]* \d+ finding.?s.? [^\d]*in )\d+\.\d+( s[\s\S]*)")
+BRANCH_COMMIT = re.compile(r"^\[(\w+) ([0-9a-f]+)\]")
+
+PIPE_OUTPUT: Mapping[str, Callable[[subprocess.CompletedProcess], str]] = {
+    "expected_out": lambda r: r.stdout,
+    "expected_err": lambda r: r.stderr,
+}
+
+
+def write_expected_file(filename: str, output: str) -> None:
+    with open(filename, "w") as file:
+        stripped = remove_commit_hash(
+            remove_trailing_space(remove_timing_seconds(output))
+        )
+        file.write(stripped)
 
 
 def remove_trailing_space(string: str) -> str:
     return "\n".join([o.rstrip() for o in string.split("\n")])
 
 
-def match_expected(output: str, expected: Expectation, test_identifier: str) -> None:
+def remove_timing_seconds(string: str) -> str:
+    result = re.sub(DYNAMIC_SECONDS, r"\1\2", string)
+    return result
+
+
+def remove_commit_hash(string: str) -> str:
+    return re.sub(BRANCH_COMMIT, r"[\1 ]", string)
+
+
+def match_expected(output: str, expected: str) -> bool:
     """Checks that OUTPUT matches EXPECTED
 
-    If EXPECTED is a string then checks that OUTPUT and EXPECTED are exact
+    Checks that OUTPUT and EXPECTED are exact
     matches ignoring trailing whitespace
 
-    If EXPECTED is a List then checks that every string in EXPECTED is contained in OUTPUT
     """
     output = remove_trailing_space(output)
 
-    if isinstance(expected, str):
-        escaped = remove_trailing_space(expected).replace("\\b", "\b")
-        if output.strip() != escaped.strip():
-            print("==== EXPECTED ====")
-            print(expected)
-            print("==== ACTUAL ====")
-            print(output.replace("\b", "\\b"))
-        assert output.strip() == escaped.strip(), test_identifier
-    else:
-        for elem in expected:
-            expected = remove_trailing_space(elem).strip()
-            if expected not in output:
-                print("==== EXPECTED ====")
-                print(expected)
-                print("==== ACTUAL ====")
-                print(output)
-            assert expected in output, test_identifier
+    # Handle dynamic characters (for now just timing info)
+    output = remove_timing_seconds(output)
+
+    if "detached HEAD" or "master" in expected:
+        output = remove_commit_hash(output)
+
+    if output.strip() != expected.strip():
+        print("==== EXPECTED ====")
+        print(expected)
+        print("==== ACTUAL ====")
+        print(output)
+    return output.strip() == expected.strip()
 
 
-def check_command(step: Any, pwd: str, target: str) -> None:
+def check_command(step: Any, pwd: str, target: str, rewrite: bool) -> None:
     """Runs COMMAND in with cwd=PWD and checks that the returncode, stdout, and stderr
     match their respective expected values.
+
+    If rewrite is True, overwrites expected files with output of running step, skipping
+    output match verification
     """
     command = step["command"]
-
-    expected_returncode = step.get("returncode")
-    expected_out = step.get("expected_out")
-    expected_err = step.get("expected_err")
-
-    # Read files if any
-    if isinstance(expected_out, dict):
-        with open(f"tests/acceptance/{target}/{expected_out['file']}") as file:
-            expected_out = file.read()
-    if isinstance(expected_err, dict):
-        with open(f"tests/acceptance/{target}/{expected_err['file']}") as file:
-            expected_err = file.read()
+    if isinstance(command, str):
+        command = command.split(" ")
 
     test_identifier = f"Target:{target} Step:{step['name']}"
     env = os.environ.copy()
     env[bento.constants.BENTO_EMAIL_VAR] = bento.constants.QA_TEST_EMAIL_ADDRESS
+    env[bento.constants.BENTO_TEST_VAR] = "1"
     substituted = [
         part.replace("__BENTO_REPO_ROOT__", BENTO_REPO_ROOT) for part in command
     ]
+
+    print(f"======= {test_identifier} ========")
 
     runned = subprocess.run(
         substituted,
@@ -82,21 +103,44 @@ def check_command(step: Any, pwd: str, target: str) -> None:
         encoding="utf-8",
     )
 
-    print(f"======= {test_identifier} ========")
     print("Command return code:", runned.returncode)
 
-    if expected_returncode is not None:
+    if "returncode" in step:
+        expected_returncode = step["returncode"]
         if runned.returncode != expected_returncode:
             print(f"Run stdout: {runned.stdout}")
             print(f"Run stderr: {runned.stderr}")
         assert runned.returncode == expected_returncode, test_identifier
-    if expected_out is not None:
-        match_expected(runned.stdout, expected_out, f"{test_identifier}: stdout")
-    if expected_err is not None:
-        match_expected(runned.stderr, expected_err, f"{test_identifier}: stderr")
+
+    for pipe in ["expected_out", "expected_err"]:
+        if pipe in step:
+            expectation_file = step.get(pipe)
+            if rewrite and expectation_file is not None:
+                write_expected_file(
+                    f"tests/acceptance/{target}/{expectation_file}",
+                    PIPE_OUTPUT[pipe](runned),
+                )
+            else:
+                if expectation_file is None:
+                    expectation = ""
+                else:
+                    with open(f"tests/acceptance/{target}/{expectation_file}") as file:
+                        expectation = file.read()
+
+                assert match_expected(
+                    PIPE_OUTPUT[pipe](runned), expectation
+                ), f"{test_identifier}: {pipe}"
 
 
-def run_repo(target: str, pre: Optional[Callable[[Path], None]] = None) -> None:
+def expand_include(step: Mapping[str, Any]) -> List[Mapping[str, Any]]:
+    include = step["include"]
+    with open(f"tests/acceptance/{include}") as file:
+        return yaml.safe_load(file)
+
+
+def run_repo(
+    target: str, pre: Optional[Callable[[Path], None]] = None, rewrite: bool = False
+) -> None:
     """
     Runs commands for a repository definition file.
 
@@ -106,57 +150,55 @@ def run_repo(target: str, pre: Optional[Callable[[Path], None]] = None) -> None:
     with open(f"tests/acceptance/{target}/commands.yaml") as file:
         info = yaml.safe_load(file)
 
-    target_repo = info["target_repo"]
-    target_hash = info["target_hash"]
+    target_repo = info.get("target_repo")
+    target_hash = info.get("target_hash")
     steps = info["steps"]
+    steps = [i for s in steps for i in (expand_include(s) if "include" in s else [s])]
 
     with tempfile.TemporaryDirectory() as target_dir:
 
-        subprocess.run(
-            ["git", "clone", target_repo, target_dir],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=True,
-        )
-        subprocess.run(
-            ["git", "checkout", target_hash],
-            cwd=target_dir,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=True,
-        )
-        subprocess.run(
-            ["git", "clean", "-xdf"],
-            cwd=target_dir,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=True,
-        )
+        if target_repo:
+            subprocess.run(
+                ["git", "clone", target_repo, target_dir],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                check=True,
+            )
+            subprocess.run(
+                ["git", "checkout", target_hash],
+                cwd=target_dir,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                check=True,
+            )
+            subprocess.run(
+                ["git", "clean", "-xdf"],
+                cwd=target_dir,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                check=True,
+            )
 
         if pre:
             pre(Path(target_dir))
 
         for step in steps:
-            check_command(step, target_dir, target)
+            check_command(step, target_dir, target, rewrite)
 
 
-# Actual Tests broken up into separate functions so progress is visible
-def test_flask() -> None:
-    run_repo("flask")
+def pytest_generate_tests(metafunc: Metafunc) -> None:
+    metafunc.parametrize("repo", ALL_TESTS)
 
 
-def test_create_react_app() -> None:
-    # eslint runs forever unless we ignore 'lib/'
-    def setup_ignores(root: Path) -> None:
-        with (root / ".gitignore").open("a") as gitignore:
-            gitignore.writelines(["lib/\n"])
-
-    run_repo("create-react-app", pre=setup_ignores)
+def test_repo(repo: str) -> None:
+    run_repo(repo)
 
 
-def test_django_example() -> None:
-    run_repo("django-example")
+if __name__ == "__main__":
+    if len(sys.argv) > 1:
+        to_run = sys.argv[1:]
+    else:
+        to_run = ALL_TESTS
 
-
-def test_instabot() -> None:
-    run_repo("instabot")
+    for t in to_run:
+        run_repo(t, rewrite=True)
